fix:refactor some of the import of types.ts
